Fixed handling of byte-order marks in input files
The new approach is to determine the presence and length of a byte-order
mark upon opening the stream. Later, when calculating byte offsets, we
just need to take into account the length of the BOM.

+++ Models/Climate/Weather.cs
@@ -962,7 +962,7 @@
         /// <param name="amp">The calculated amp value</param>
         private void ProcessMonthlyTAVAMP(out double tav, out double amp)
         {
-            int savedPosition = reader.GetCurrentPosition();
+            long savedPosition = reader.GetCurrentPosition();
 
             // init return values
             tav = 0;
